% Generated by roxygen2 (4.0.2): do not edit by hand
\name{checkPolyHole}
\alias{checkPolyHole}
\title{Internal function of spToGeoJSON}
\usage{
checkPolyHole(polyCoord, holeCoord)
}
\arguments{
\item{polyCoord}{Polygon coordinates as data.frame.}

\item{holeCoord}{Hole coordinates as data.frame.}
}
\value{
Boolean
}
\description{
Checks if hole is part of polygon or not. Used by \code{\link{spToGeoJSON}}.
}
\author{
Christian Graul
}
\seealso{
\code{\link{spToGeoJSON}}
}
\keyword{internal}
\keyword{methods}

