% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getFeatureType}
\alias{getFeatureType}
\title{Internal function}
\usage{
getFeatureType(data)
}
\arguments{
\item{data}{GeoJSON or TopoJSON object.}
}
\value{
FeatureType as string.
}
\description{
Checks for type of features stored in a GeoJSON file.
}
\author{
Christian Graul
}
\keyword{internal}
\keyword{methods}

