% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getHex}
\alias{getHex}
\title{Internal function for color conversion}
\usage{
getHex(col)
}
\arguments{
\item{col}{Color value as number, name or hex code with alpha.}
}
\value{
Hexadecimal color value as string.
}
\description{
Converts different color values to hexadecimal colors. Used by
\code{\link{styleSingle}}, \code{\link{styleGrad}} and
\code{\link{styleCat}}.
}
\author{
Christian Graul
}
\seealso{
\code{\link{styleSingle}}, \code{\link{styleGrad}},
\code{\link{styleCat}}
}
\keyword{internal}
\keyword{methods}

