% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getProperties}
\alias{getProperties}
\alias{prop}
\title{Print property names of a GeoJSON file}
\usage{
getProperties(data, print = TRUE)
}
\arguments{
\item{data}{Path to data file as string or GeoJSON/TopoJSON object.}

\item{print}{If \code{TRUE} (default), the property names are printed.}
}
\value{
Property names as string vector.
}
\description{
Prints the names of all available properties of a GeoJSON file.
}
\examples{
\dontrun{
data(quakes)
qks <- toGeoJSON(data=quakes, dest=tempdir())
getProperties(data=qks)
}
}
\author{
Christian Graul
}
\seealso{
\code{\link{styleCat}}, \code{\link{styleGrad}}
}
\keyword{methods}

