% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getTopologies}
\alias{getTopologies}
\alias{topo}
\title{Print topology objects of a TopoJSON file}
\usage{
getTopologies(data, print = TRUE)
}
\arguments{
\item{data}{Name of data file as string or TopoJSON object.}

\item{print}{If \code{TRUE} (default), the property names are printed.}
}
\value{
Topology object names as string vector.
}
\description{
Prints the names of all available topology objects of a TopoJSON file.
}
\examples{
\dontrun{
#getTopologies()
}
}
\author{
Christian Graul
}
\keyword{methods}

