% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fileToGeoJSON}
\alias{fileToGeoJSON}
\title{Internal function of toGeoJSON}
\usage{
fileToGeoJSON(data, name, dest, overwrite)
}
\arguments{
\item{data}{Path to external spatial data file, as string.}

\item{name}{Name of the resulting GeoJSON file, as string.}

\item{dest}{Directory the file shall be saved to, as string.}

\item{overwrite}{\code{TRUE} overwrites existing files with the same name.}
}
\value{
GeoJSON file path, as string.
}
\description{
Create GeoJSON file from external spatial data file
}
\author{
Christian Graul
}
\seealso{
\code{\link{toGeoJSON}}
}
\keyword{internal}
\keyword{methods}

