\encoding{UTF-8}
\name{getFeatureType}
\alias{getFeatureType}
\title{Internal function}
\description{Checks for type of features stored in a GeoJSON file.}
\usage{getFeatureType(dat)}
\arguments{
  \item{dat}{Name of data file as string.}
}
\value{FeatureType as string.}
\author{Christian Graul}
\keyword{methods}
\keyword{internal}