\encoding{UTF-8}
\name{getFeatureType}
\alias{getFeatureType}
\title{Internal function}
\description{Checks for type of features stored in a GeoJSON file.}
\usage{getFeatureType(data)}
\arguments{
  \item{data}{GeoJSON or TopoJSON object.}
}
\value{FeatureType as string.}
\author{Christian Graul}
\keyword{methods}
\keyword{internal}