\encoding{UTF-8}
\name{getProperties}
\alias{getProperties}
\title{Internal function}
\description{Get the names of all properties of a GeoJSON file. Used by \code{\link{leaflet}}.}
\usage{getProperties(dat)}
\arguments{
  \item{dat}{Name of data file as string.}
}
\value{Properties as string vector.}
\author{Christian Graul}
\seealso{\code{\link{leaflet}}}
\keyword{methods}
\keyword{internal}